+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Add support for `false` booleans to `getFieldValue` util (issue #1)
+
 ## v0.1.1 (October 09, 2023)
 
 - Improve text and change example in README.md